% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateHelper.R
\name{dateHelper}
\alias{dateHelper}
\title{Robust conversion of many date/time formats to Date}
\usage{
dateHelper(x)
}
\arguments{
\item{x}{character/numeric/Date/POSIXt}
}
\value{
Date
}
\description{
Handles Excel serials, UNIX epoch (sec/ms), and common string formats
like "11/16/2018 0:00", "2024-05-12", "2/20/2001 12:34 PM".
Returns Date (YYYY-MM-DD), dropping time.
}
